Develop

Closes SFDM-2874 and SFDM-2834

See merge request products/eafdmmart/fdm-products!69

+++ src/main/java/ru/beeline/fdmproducts/domain/Sla.java
@@ -30,5 +30,5 @@ public class Sla {
     private Integer latency;
 
     @Column(name = "error_rate")
-    private Integer errorRate;
+    private Double errorRate;
 }
